Revert input handling implementation to the previous scheme, where each frame will wait for a user input before proceeding.
But this is not completely the same as the previous scheme because it contains
fixes for bugs that was found in the previous implementation, One main bug in
the previous implementation was how the input was only sent on HandleInput and
not on GrabInput. This meant that the inputs were not being pipelined through
the network in any way so the FRAME_GAP wasn't actually being used in the way it
was intended to.

This commit seems to fix the input handling scheme in the I-Frame technique
properly, where we can achieve both good FPS and good input latencies